自定义键盘，系统拍照工具类的实现、去掉use_frameworks!以及由此引起的错误的修改

+++ XMPP_TestDemo/XMPP/ChatMVC/View/XZTextView.m
@@ -28,7 +28,7 @@
 {
     self.placeholderView.hidden = self.hasText;
     
-    NSInteger height = ceilf([self sizeThatFits:CGSizeMake(self.bounds.size.width, MAXFLOAT)].height);
+    CGFloat height = ceilf([self sizeThatFits:CGSizeMake(self.bounds.size.width, MAXFLOAT)].height);
     
 //    if (_textH != height) { // 高度不一样，就改变了高度
